Hash: add hardware hashing solutions to the hash command
Now the command provides the interface to access the driver
of the hash accelerator.

In order to access the hardware hashing methods, one of the
following algorithm names must be used:
hwsha1, hwsha256, hwsha224, hwmd5

The command should have the following structure:
"hash" algo_name img_addr img_length

img_addr and img_length is in hexadecimal and digest_size is in decimal.

An example command is:
"hash hwsha1 0xBA000050 C"

Change-Id: I81681913d6053c4ffda69126af9e0d76012132b1
Signed-off-by: Tianci Ma <Tianci.Ma@imgtec.com>