mtd: nand: Allow to set a per-device ECC layout
Until now, the ECC was handled either by software or by
the NAND controller (hardware ECC). In each of these, the
ECC layout is defined by the controller of the NAND device,
not by the NAND device itself.

However, devices with on-die ECC support have their own
ECC layout so we need a way to specify a per-device ECC layout.

This commit adds a new field to the nand_flash_dev structure,
to allow devices to specify its own ECC layout.

Change-Id: I3ddb53afcbe3843cd2a7120b4e59e41fb6222860
Author: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>