pistachio: Ethernet: allow mac address to be parsed from dtb
Currently a default mac addr is passed from board file
remove this infrastructure and use mac address from dtb.
To parse the mac-addr add a alias node to ethernet,
If mac if parsed from dtb from u-boot it also passes this info
to kernel by doing a dtb fixup, so add this alias node
which is used for mac-addr fix-up to kernel and used
to derive the mac-addr before device init.

Change-Id: I7188b93bd5962574a8a4768bc06427c99a155d5e
Signed-off-by: Govindraj Raja <Govindraj.Raja@imgtec.com>