nor: add infrastructure for pistachio NOR images
Using the new added Makefile targets an image having the
u-boot SPL, DTBs and the u-boot payload can be created.

TODO: simplify the creation of the binary holding the
SPL and DTB. Given that we have a large GRAM the BSS
area can be continuous to the text and data segments
and the DTB can be placed imediatelly after BSS.
The image in this case can be simply added with "cat".

Change-Id: I1273d17a8746c8598eae2596550765165ef295a9
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>