SPL: place DTB imediatelly after text and data segments
This makes the procedure to obtain the SPL+DTB easier as
all we have to do is cat the two images together, without
having to specify an offset for DTB inside the final SPL
image (no dd required).
Having done this requires moving the BSS to SRAM, which is
also used for the stack.

Change-Id: I603ef59c396a1e30624480949fad040b3ef81839
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>