mips: mmu: make text start and size 8KB aligned
Each TLB entry maps a pair of virtual pages of minimum 4KB.
Therefore both text start and text size (and in consequence
data_start) has to be 8KB aligned.

Without this change the addition of the hash accelerator will
determine the increase of the current text size which will
in turn trigger an exception when this alligment is not
respected.

Change-Id: Ib2cc8ad21011f169c14dea01c109462eda914b8f
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>