I2C: IMGTEC driver implementation
This is a driver for the IMG serial control bus (SCB) IP block.
The driver is a simplified version of the Linux driver for this
block, having the I2C controller driven in automatic mode, with
state control guided by polling the interrupt registers.

In automatic mode, a bus address, a data count and whether to
read or write is specified, and the hardware takes care of the
I2C state machine, using a FIFO to send/receive bytes of data
to/from an I2C slave. The driver has to keep the FIFO drained
or filled in response to the values of the interrupt registers.

Change-Id: Ia094672c647ad5e4427672fc1c7c4022d4ff1473
Signed-off-by: Tianci Ma <Tianci.Ma@imgtec.com>