pistachio: Disable default RPU gate reg values.
The RPU Clock register defaults to ON for all clocks,
this should be modified to OFF, and the MIPS Clk control
modified to ON by default. This is because the Kernel will
manage the clocks at all times, but the RPU can only disable
clocks if the WIFI module has been loaded (and correctly coded).

Bug-ID: 74172
Change-Id: I339f91cf1cf4d5209c5d3f09520a665ac5391475
Signed-off-by: Govindraj Raja <Govindraj.Raja@imgtec.com>