IMG SPFI: add support for dual and quad mode
Information about the RX/TX bus widths will be passed from
the flash device driver through spi_slave.

For dual and quad modes to work the SPFI controller needs
to have information about command/address/dummy bytes in the
transaction register. This information is not relevant for
single mode, and therefore it can have any value in the
allowed range. Therefore, for any read or write transfers of less
than 8 bytes (cmd = 1 byte, addr up to 7 bytes), SPFI will be
configured, but not enabled (unless it is the last transfer in
the queue). The transfer will be enabled by the subsequent transfer.
A pending transfer is determined by the content of the transaction
register: if the command part is set and tsize is not.
This way we ensure that for dual and quad transactions
the command request size will appear in the command/address part
of the transaction register, while the data size will be in
tsize, all data being sent/received in the same transaction (as
set up in the transaction register).

Change-Id: I88d04faa70654ed66ce7d44829007f4dd22d4300
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>