IMG SPFI: optimize transfer function
For each byte/4-bytes transferred the get_timer will be called.
That function is very time consuming as it does the division
between the number of ticks and the CPU frequency by repeated
substraction.
Therefore, optimize the SPI transfer function to only call
get_timer when there were no bytes transferred in the current iteration.
If the SPFI HW fails and hangs, it won't output/input new bytes,
so no bytes will be transferred, and the deadline will be exceeded.

Test: the time to read 128MB is reduced from 4m:23s to 1m:01s.

Change-Id: I966db069ee47c4c0ef1bbcaa6cc37d5619e9ef8f
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>