pistachio: fix mmc bus frequency
Currently the mmc/SDHost device is performing slower
than it should be because the block is configured for a 
200MHz input clock, but the actual input clock is 87.5MHz.
The mmc/SDHost block has a fixed divide by 4 from the system 
clk (350MHz/4 = 87.5MHz), so update the device configuration 
to reflect this.

When the driver calculates the frequency needed for the 
card inserted, it will now be as close as possible to the 
correct frequency.  For example, an SD card that previously
was clocked at 21.875MHz will now be clocked at 43.75MHz

Change-Id: Ieb5989c7b479d19c492f07cf02f1f6d6b20f9610
Signed-off-by: Govindraj Raja <Govindraj.Raja@imgtec.com>