mtd: spi-nand: Implement quad-mode for transmit
Until now, only single mode operations were supported for TX
transfers. With this change, store cache operations will,
from this moment on, be performed in either quad mode or
single mode, depending on the bus width set in device tree.

This requires changes to the SPFI driver consisting of setting
the TX_RX bit which lets the state machine know that the command
which is going to be sent to the flash is a TX (set)/RX(cleared)
command. This helps with the scenario in which we have both a
command transfer (command, address, dummy) which will be pending(TX),
and a TX data transfer, which was problematic before. This bit of
the control register is flash specific and has no significance in
SPI traditional sense.

Change-Id: Iceedd0a3438f78184d052e0b972de7478e3856f3
Signed-off-by: Ionela Voinescu <ionela.voinescu@imgtec.com>